split build tests into subdirectories
