+++ uniplate/tests/derive-pass/enums/allow-enum-fields-with-trailing-commas.rs
