+++ uniplate/tests/derive-pass/enums/allow-enum-variant-with-no-fields.rs
