+++ uniplate/tests/derive-pass/enums/allow-enums-with-trailing-commas.rs
