+++ uniplate/tests/derive-pass/structs/allow-tuple-struct.rs
@@ -1,0 +1,7 @@
+use uniplate::derive::Uniplate;
+
+#[derive(Uniplate, PartialEq, Eq, Clone)]
+#[uniplate()]
+struct Tuple(i32, String);
+
+fn main() {}
