+++ uniplate/tests/derive-pass/structs/allow-unit-like.rs
@@ -1,0 +1,7 @@
+use uniplate::derive::Uniplate;
+
+#[derive(Uniplate, PartialEq, Eq, Clone)]
+#[uniplate()]
+struct UnitLike;
+
+fn main() {}
