+++ uniplate/tests/derive-pass/structs/recursive-structs.rs
@@ -1,0 +1,10 @@
+use uniplate::derive::Uniplate;
+
+#[derive(Uniplate, PartialEq, Eq, Clone)]
+#[uniplate()]
+struct Tree {
+    value: i32,
+    children: Vec<Tree>,
+}
+
+fn main() {}
