+++ uniplate/tests/derive_test_runner.rs
@@ -1,7 +1,7 @@
 #[test]
 fn derive_tests() {
     let t = trybuild::TestCases::new();
-    t.compile_fail("tests/derive-fail/*.rs");
-    t.pass("tests/derive-pass/*.rs");
-    t.pass("examples/*.rs");
+    t.compile_fail("tests/derive-fail/**/*.rs");
+    t.pass("tests/derive-pass/**/*.rs");
+    t.pass("examples/**/*.rs");
 }
